requirements.txt
opencv-python
pillow
deepface
numpy
sqlite3
